perf(test): SDK型定義バンドルの重さを調査・分離（タイムアウト延長ではなく根本対応）
「タイムアウトを伸ばすのでなく軽くできないか」という指摘を受け、buildDts(dts-bundle-generator)
の21秒がどこから来るか二分探索で実測した。

## 調査結果
- zodスキーマの混入が原因という当初の仮説は誤り。@ubichill/shared単体・@ubichill/ecs単体・
  UbiSDKクラス単体はいずれも数百ms〜1秒で、個別には重くない。
- 本当の原因: dts-bundle-generatorがSDKの公開API全体（100件超の再エクスポート型）を
  1プログラムとして型チェックする際のTypeScriptチェッカー側の非線形コスト。多数の複雑な型を
  組み合わせるほど不均衡に遅くなる（個々のサブパスは高速でも、全部combineすると21秒）。
- noCheckオプション・shared barrel分離・webworker lib絞り込みなど複数の軽減策を実測したが、
  いずれも有意な改善は得られなかった（21.7秒→20.9秒、誤差程度）。

## それでも実施した正当な改善（副産物）
- packages/shared/src/mod/entities.ts 新規: zod非依存の素朴な型（User/ComponentInstance等）を
  barrel（index.ts、schemas/*のzodスキーマを抱える）から分離。mod/types.tsがindex.tsを逆参照
  する循環依存を修正（entities.ts参照に変更）。sharedのpackage.jsonにサブパスexport追加
  （./mod/protocol, ./mod/errors, ./mod/entities）。
- packages/sdk/src/**: @ubichill/sharedのbarrel importを全廃し、実際に使うサブパス
  （/mod/entities, /mod/types, /mod/protocol, /mod/errors, /mod/vnode）を直接importに変更。
  責務がより明確になった（型の出自が一目で分かる）。
- packages/sdk/tsconfig.json: SDKはWorker実行なのでlibをdom/dom.iterableからwebworkerに
  訂正（より正確）。未使用だったsandbox/react/ui-renderer/shared(bare)のpathsを削除。

## テストの分離（本題の対応）
buildDts自体は速くできなかったため、通常の開発ループを重くしないよう分離した:
- scripts/build-sdk.test.mjs: 高速なチェックのみ残す（buildJs, buildPackageJson, LICENSE存在）。
  `pnpm test`は変わらず約1秒。
- scripts/build-sdk.dts.verify.mjs 新規: buildDts依存の重いテスト（型定義の外部参照ゼロ検証、
  未インストール環境でのtsc統合検証）を分離。ファイル名を意図的に*.test.mjsにせず、既定の
  vitest includeから除外。`pnpm verify:sdk-types`（専用configで明示実行）でCIから呼ぶ。
- .github/workflows/ci.yml: test jobに`pnpm verify:sdk-types`ステップを追加（検証は失わない）。
- vitest.config.ts: alias指定をplain objectからregex-anchored arrayに変更（prefixマッチで
  サブパスimportが壊れるバグを発見・修正。@ubichill/shared/mod/*等の新サブパスaliasを追加）。

## 検証
- pnpm test（既定）: 213 tests pass、約1秒（従来と同等の速さに復帰）。
- pnpm verify:sdk-types: 4 tests pass、約22秒（buildDts自体の不可避なコスト、分離済みなので
  通常ループには影響しない）。
- pnpm typecheck / pnpm lint: 全体クリーン。
- pnpm build:workers: hash不変（回帰なし）。pnpm build:sdk: 正常にdist-npm生成。
