+++ .github/workflows/ci.yml
@@ -111,6 +111,12 @@ jobs:
       - name: Verify mod locks
         run: node scripts/verify-mod-locks.mjs
 
+      # SDK型定義バンドルの検証（buildDts = TS Compiler API フルコンパイル、実測20秒級）。
+      # 開発ループを重くしないためファイル名を *.test.mjs にせず既定のvitest includeから
+      # 除外している（scripts/build-sdk.dts.verify.mjs 冒頭コメント参照）。CIでは明示的に実行する。
+      - name: Verify SDK type bundle
+        run: pnpm verify:sdk-types
+
   # 変更検知: PR 時はパスフィルタ、push to main では全て true を返す。
   detect-changes:
     name: Detect changed packages
